these are the changes I have made to our Task Manager Project.
I made a table view to hold all the tasks with a smaple row at the top. The first text field displays the task name and the second displays the time and the last one shows if the task has been completed or not. The add button shows a new view controller the allows you to input the data for a new task.